Swallow fixups. Had merged MEA cmake into this branch, overzealous with a few of the manual confict fixups (favoured MEA each time). Looks like there are some subtle differences to Swallow types that use variables that were otherwise unused in MEA. Also generated CMakeLists for files added to Swallow that aren't in MEA. Fixed up plist file template to match Swallow's existing version. Minor fixes to warnings found in Swallow only files (already done for MEA). Buddy: Vimmy

+++ src/SwallowPoiDb/Parsers/SwallowPoiDbParseHelper.cpp
@@ -17,7 +17,7 @@ namespace ExampleApp
 			{
 				rapidjson::Value resultValue(rapidjson::kStringType);
 
-				resultValue.SetString(value.c_str(), value.size(), allocator);
+				resultValue.SetString(value.c_str(), static_cast<int>(value.size()), allocator);
 
 				out_jsonValue.AddMember(name.c_str(),
 					resultValue,
